Move the explore map to its own component
